Move createThrowsSpec to ElementHandlerUtil
